Added installation progess to console output

+++ Install-NgOIOUBLMover.ps1
@@ -227,30 +227,39 @@ function Install-NgFiles {
     }
     else{
         try {
+            write-host "Checking connection to '$($GitHubRepoUrl.Uri)'" -ForegroundColor Yellow
             $HTTP_ResponseRepo = $HTTP_RequestRepo.GetResponse()
+            write-host "Success: Connection was successful" -ForegroundColor Green
         }
         catch {
-            write-NgLogMessage -Message "Unable to connect to $GitHubRawUrl" -Level Error
-            write-Error "Install-NgFiles: Unable to connect to $GitHubRawUrl $_"
+            write-host "Error: Connection failed`nUnable to connect to '$($GitHubRepoUrl.Uri)'`nTerminating installation" -ForegroundColor Red
+            write-NgLogMessage -Message "Unable to connect to '$($GitHubRepoUrl.Uri)'" -Level Error
+            write-Error "Install-NgFiles: Unable to connect to '$($GitHubRepoUrl.Uri)' $_"
             throw $_
             exit 1
         }
         
         if ([int]$HTTP_ResponseRepo.StatusCode -ne 200){
+            write-host "Error: Connection failed`nUnable to connect to '$($GitHubRepoUrl.Uri)'`nTerminating installation" -ForegroundColor Red
             write-NgLogMessage -Message "Unable to download files from $GitHubRepoUrl" -Level Error
             exit 1
         }
         if ($Force){
             try {
-                Remove-Item $InstallPath -Recurse -Force
+                if (Test-Path $InstallPath){Remove-Item $InstallPath -Recurse -Force}
+                
+                Write-host "Downloading source files from '$($GitHubRawUrl.Uri)'" -ForegroundColor yellow
+
                 Invoke-WebRequest -Uri $GitHubRawUrl.Uri -OutFile (Join-Path -Path $InstallLocation -ChildPath "NgEANMover.Zip")
                 Expand-Archive (Join-Path -Path $InstallLocation -ChildPath "NgEANMover.Zip") $InstallLocation
                 Move-Item (Join-Path -Path $InstallLocation -ChildPath "NgEANMover-main") $InstallPath -Force
                 Remove-Item (Join-Path -Path $InstallLocation -ChildPath "NgEANMover.Zip")
-
+                
+                write-host "Success: Downloaded and extracted source files to '$($InstallPath)'" -ForegroundColor Green
                 write-NgLogMessage -Message "Downloaded $($GitHubRawUrl.Uri) to $(Join-Path -Path $InstallPath -ChildPath 'NgEANMover.Zip')" -Level Information
             }
             catch {
+                write-host "Error: Failed to download source files`nTerminating installation" -ForegroundColor Red
                 write-NgLogMessage -Message "Unable to download $($GitHubRawUrl.Uri) to $InstallPath $_" -Level Error
                 Write-Error "Install-NgFiles: Unable to download $($GitHubRawUrl.Uri) to $InstallPath $_"
                 return $_
@@ -260,11 +269,13 @@ function Install-NgFiles {
         else {
             foreach ($MissingFile in $MissingFiles) {
                 try {
+                    Write-host "Downloading source files from '$($GitHubRawUrl.Uri)'" -ForegroundColor yellow
                     Invoke-WebRequest -Uri "$($GitHubRawUrl.Uri)/$MissingFile" -OutFile (Join-Path -Path $InstallPath -ChildPath $MissingFile)
-                    
+                    write-host "Success: Downloaded source files to '$($InstallPath)'" -ForegroundColor Green
                     write-NgLogMessage -Message "Downloaded $MissingFile to $InstallPath" -Level Information
                 }
                 catch {
+                    write-host "Error: Failed to download source files`nTerminating installation" -ForegroundColor Red
                     write-NgLogMessage -Message "Unable to download $MissingFile to $InstallPath $_" -Level Error
                     Write-Error "Install-NgFiles: Unable to download $MissingFile to $InstallPath $_"
                     return $_
@@ -274,22 +285,28 @@ function Install-NgFiles {
         
         if ($Type -eq "exe") {
             try {
+                write-host "Installing ps2exe module" -ForegroundColor Yellow
                 Set-PSRepository "PSGallery" -InstallationPolicy Trusted
                 Install-Module ps2exe -Scope CurrentUser -Force -AllowClobber
                 Import-Module ps2exe -Force
+                write-host "Success: installed ps2exe" -ForegroundColor Green
                 write-NgLogMessage -Message "Installed ps2exe module" -Level Information
             }
             catch {
+                write-host "Error: Unable to install ps2exe module" -ForegroundColor Red
                 write-NgLogMessage -Message "Unable to install ps2exe module $_" -Level Error
                 Write-Error "Install-NgFiles: Unable to install ps2exe module $_"
                 return $_
             }
     
             try {
+                write-host "ps2exe will now complice source files to EXE" -ForegroundColor Yellow
                 Invoke-ps2exe -inputFile $Compile -outputFile "$($Compile -replace '.ps1','.exe')" -title "EAN Mover" -company "NgMS Consult ApS" -version "2.0" -product "NgOIOUBLMover" -noConsole -copyright "Copyright (c) 2024 - Phillip Schjeldal Hansen | NgMS Consult ApS. All rights reserved." -longPaths -iconFile (Join-Path -Path $InstallPath -ChildPath $Icon) -configFile
+                Write-Host "Success: Compiled NgOIOUBLMover.exe" -ForegroundColor Green
                 write-NgLogMessage -Message "Created executable NgOIOUBLMover.exe" -Level Information
             }
             catch {
+                write-host "Error: Failed to compile NgOIOUBLMover.exe`nTerminating installation" -ForegroundColor Red
                 write-NgLogMessage -Message "Unable to create executable $_" -Level Error
                 Write-Error "Install-NgFiles: Unable to create executable $_"
                 return $_
@@ -297,6 +314,12 @@ function Install-NgFiles {
         }
     }
 }
+Write-Host "##############################################" 
+Write-Host "Starting installation of EAN Mover" -ForegroundColor Green
+Write-Host "COPYRIGHT (c) 2024 - Phillip Schjeldal Hansen | NgMS Consult ApS. All rights reserved" -ForegroundColor Green
+Write-Host "##############################################" 
+Write-Host ""
+
 
 $TaskName = "NgOIOUBLMover"
 $TaskDescription = "Move OIOUBL/EAN files from the downloads folder to $AzureFileShare"
@@ -314,11 +337,15 @@ $LogPath = $env:temp
 [string]$LogFilePrefix = "Install_" # Date will be appended to the prefix ex. Install_10-12-2024.log
 
 try {
+
+    Write-Host "Checking if EAN Mover is currently running, if already installed" -ForegroundColor Yellow
     #check if NgOIOUBLMover is already running
     if ((Get-Process -Name "NgOIOUBLMover" -ErrorAction SilentlyContinue).count -gt 0) {
+        Write-Host "EAN Mover already running`nPlease wait for it to complete and start the installation again`nTerminating installation" -ForegroundColor Red
         [System.Windows.Forms.MessageBox]::Show($THIS, "EAN Mover already running`nPlease wait for it to complete and start the installation again",'EAN Mover','OK','error')
         exit "Process already running"
     }
+    write-host "Success: EAN Mover is not running`nInstallation will continue" -ForegroundColor Green
 
     # Set the log folder
     [string]$LogFolder = Join-Path -Path $LogPath -ChildPath $FolderName # Log files will be stored in the temp folder in a folder named NgOIOUBLMover
@@ -347,14 +374,23 @@ try {
 
     if (-not $DisableStartMenuShortcut) {
         Add-NgStartMenuShortcut -ScriptPath $NgScriptPath -ScriptParameters $NgScriptParameters
+        write-host "Success: Created Start menu shortcut" -ForegroundColor Green
     }
 
     if (-not $DisableDesktopShortcut) {
         Add-NgDesktopShortcut -ScriptPath $NgScriptPath -ScriptParameters $NgScriptParameters
+        write-host "Success: Created desktop shortcut" -ForegroundColor Green
     }
 
     Add-NgScheduledTask -TaskName $TaskName -ScriptPath $NgScriptPath -ScriptParameters $NgScriptParameters -TaskDescription $TaskDescription -TaskInterval $RunInterval -TaskId 124563 -Disabled $DisableScheduledTask -TimeOut 15
+    write-host "Success: Created Scheduled Task" -ForegroundColor Green
     write-NgLogMessage -Message "Installation of NgOIOUBLMover completed" -Level Information
+    Write-Host ""
+    Write-Host "##############################################" 
+    Write-Host "Success: Finished installation of EAN Mover" -ForegroundColor Green
+    Write-Host "##############################################"
+
+
     [System.Windows.Forms.MessageBox]::Show($THIS, "Installation of NgOIOUBLMover Completed",'OIOUBL Mover','OK','Information')
 }
 catch {
